+++ Models/AudioInsightSeverity.cs
@@ -1,0 +1,8 @@
+namespace AudioQualityEnhancer.Models;
+
+public enum AudioInsightSeverity
+{
+    Info,
+    Warning,
+    Critical
+}